% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypeR.R
\docType{package}
\name{clonotypeR}
\alias{clonotypeR}
\alias{clonotypeR-package}
\title{clonotypeR: High throughput analysis of T cell antigen receptor sequences}
\description{
clonotypeR: High throughput analysis of T cell antigen receptor sequences
}

