\name{J_before_FGxG}
\alias{J_before_FGxG}
\docType{data}

\title{Sequence of the J segments before their conserved FGxG motif.}

\description{
Table intended for internal use in the \code{yassai_identifier}
function, generated from the reference data with the
command: \code{make refresh-data}.
}

\usage{
data(J_before_FGxG)
}

\value{
\item{row name}{J segment name, for instance \dQuote{TRAJ61}.}
\item{sequence}{Sequence of the nucleic acids preceding the first codon of the conserved FGxG motif.}
}

\seealso{
\code{\link{yassai_identifier}}
}

\keyword{ clonotype }
\keyword{ conserved cysteine }
\keyword{ J segment }
