\name{V_after_C}
\alias{V_after_C}
\docType{data}

\title{Sequence of the V segments after their conserved cystein.}

\description{
Table intended for internal use in the \code{yassai.identifier}
function, generated from the reference data with the
command: \code{make refresh-data}.
}

\usage{
data(V_after_C)
}

\value{
\item{row name}{V segment name, for instance \dQuote{TRAV1}.}
\item{sequence}{Sequence of the nucleic acids following the codon of the conserved cysteine.}
}

\seealso{
\code{\link{yassai.identifier}}
}

\keyword{ clonotype }
\keyword{ conserved cysteine }
\keyword{ V segment }
