% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_clonotypes.R
\name{private_clonotypes}
\alias{private_clonotypes}
\title{private_clonotypes}
\usage{
private_clonotypes(..., data)
}
\arguments{
\item{...}{Library names.}

\item{data}{A clonotype table.}
}
\value{
A vector of clonotype names.
}
\description{
List clonotypes found exclusively in one library.
}
\examples{
clonotypes <- read_clonotypes(system.file('extdata', 'clonotypes.txt.gz', package = "clonotypeR"))
clonotypes <- clonotype_table(levels(clonotypes$lib), data=clonotypes)
private_clonotypes("C", data=clonotypes)
}
\seealso{
\code{\link{clonotype_table}}
}

